Remove deref to frontend and store value and original
Storing the value and the original cuts out repeated conversions making
change operations much faster and thinner on top of the frontend change.

Deref makes it likely that users will cause the stored values to get out
of sync with the real values unless they manually refresh them. Removing
those and implementing the aspects manually is a better approach.